Revert "add 3.7 Kinesis integration"
This reverts commit ce5d382a856112e42f50d5f95c3f5f4bb1902857.
