+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AWS/Kinesis/IPutRecordRequest.cs
@@ -10,7 +10,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AWS.Kinesis
     /// <summary>
     /// PutRecordRequest interface for duck typing.
     /// </summary>
-    internal interface IPutRecordRequest : IAmazonKinesisRequestWithStreamName, IContainsData
+    internal interface IPutRecordRequest : IAmazonKinesisRequest, IContainsData
     {
     }
 }
