+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AWS/Kinesis/IPutRecordsRequest.cs
@@ -12,7 +12,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AWS.Kinesis
     /// <summary>
     /// PutRecordsRequest interface for duck typing.
     /// </summary>
-    internal interface IPutRecordsRequest : IAmazonKinesisRequestWithStreamName
+    internal interface IPutRecordsRequest : IAmazonKinesisRequest
     {
         /// <summary>
         /// Gets or sets the Kinesis Records.
