--- tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AWS/Kinesis/IPutRecordsRequestV3_7.cs
@@ -1,22 +1,0 @@
-// <copyright file="IPutRecordsRequestV3_7.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-#nullable enable
-
-using System.Collections;
-
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AWS.Kinesis
-{
-    /// <summary>
-    /// PutRecordsRequest interface for duck typing.
-    /// </summary>
-    internal interface IPutRecordsRequestV3_7 : IAmazonKinesisRequestWithStreamNameAndStreamArn
-    {
-        /// <summary>
-        /// Gets or sets the Kinesis Records.
-        /// </summary>
-        IList? Records { get; set; }
-    }
-}
